being processing of splitting out draft service responsbilities
